+++ src/Glutinum.Converter/GlueAST.fs
@@ -60,12 +60,20 @@ type GlueIndexSignature =
         Type: GlueType
     }
 
+type GlueMethodSignature =
+    {
+        Name: string
+        Parameters: GlueParameter list
+        Type: GlueType
+    }
+
 [<RequireQualifiedAccess>]
 type GlueMember =
     | Method of GlueMethod
     | Property of GlueProperty
     | CallSignature of GlueCallSignature
     | IndexSignature of GlueIndexSignature
+    | MethodSignature of GlueMethodSignature
 
 type GlueInterface =
     {