+++ tests/specs/interfaces/method.d.ts
@@ -1,0 +1,3 @@
+interface Locale {
+    hello(): string;
+}