+++ tests/specs/keyof/method.d.ts
@@ -1,0 +1,5 @@
+export interface Logger {
+    hello() : string;
+}
+
+type L = keyof Logger;